Minor: delete unnecessary call to make_seed_mech outside RMG while loop
make_seed_mech() is called at the start of each iteration just after entering the main RMG while loop in execute, which saves seed mechanism. If more iterations must be done (due to species being added to the core or additional reactions being added to the edge etc), it returns to the top of the while loop and saves the new seed mechanism. If no changes were made, it exits the while loop and previously proceeded to call make_seed_mech() again, which is unnecessary because there are no changes.  @amarkpayne and I verified that every file contained within the filters, seed, and seed_edge folders were identical regardless of whether this line was present. Thus, there is no need to call make_seed_mech() again