Refactor writing of elements section of Chemkin thermo entries
For species with 4 elements or less, put them in the first line
For species with 5 elements or more, put them on the next line
using the space delimited syntax.

Although the CHEMKIN manual specifies that one may put
the 5th element atom count in columns 74-78 of row 1,
in practice nobody does this, and Cantera cannot read it
if we do, so this commit stops us from doing this.
Instead if there are 5 or more elements present, we use
the extended syntax (that we were using for 6+ elements).

For more details see discussion at
https://github.com/Cantera/cantera/issues/656